Resolve S09P12A308-236 "Feat/be 회원관리 회원가입 시 전체 언어 등급 리스트 조회"

Closes S09P12A308-236

See merge request s09-webmobile1-sub2/AirLingo!14

+++ backend/airlingo/src/main/java/com/ssafy/airlingo/domain/language/dto/request/EvaluateUserRequestDto.java
@@ -19,10 +19,10 @@ public class EvaluateUserRequestDto {
 	private Long userId;
 
 	@Schema(description = "언어 등급 평가 결과 ID", example = "2")
-	private Integer gradeId;
+	private Long gradeId;
 
 	@Schema(description = "사용자가 사용한 언어 ID", example = "1")
-	private Integer languageId;
+	private Long languageId;
 
 	@Schema(description = "스터디 ID", example = "1")
 	private Long studyId;
